+++ package.json
@@ -28,7 +28,8 @@
     "naive-ui": "^2.34.4",
     "pinia": "^2.1.4",
     "vue": "^3.3.4",
-    "vue-material-design-icons": "^5.2.0"
+    "vue-material-design-icons": "^5.2.0",
+    "vue-router": "4"
   },
   "devDependencies": {
     "@typescript-eslint/eslint-plugin": "^5.60.1",
